dataV2/1_sql/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0